requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
loguru==0.6.0
ujson==5.7.0
redis==4.4.0
requests==2.28.1
geographiclib==2.0
geopy==2.3.0
yahoo-fin==0.8.9.1
telebot==0.0.5
selenium==4.18.1